feat(Epic 4 User Account Delete):
User can delete their account,
links to the delete profile page are added to the dashboard
and profile details.
The account is deleted, and orders remain with the delivery details etc.
If user creates a new account the same user details can be used, but
the previous orders are linked to the original PK not the user name.
They are linked this way for security incase sombody else uses the same
username after the original profile is deleted.