feat(Epic 7 Product Creation/Edit):
Product creation and editing from the front end.
menu item added to menu visable to staff only for create new product.
Icon next to product name as link to edit product, visable to staff only
when new image is uploaded the current image is replace as preview in
both the create and edit views.