feat(Epic 7 Product Deletion):
Added an icon to delete the product.
This opens a modal to delete the product.
Modal advises to set product as inactive and gives a link to  edit
the product and a cancel button which returns to the original product.
Delete will delete the product and return to the product list