feat: added pagination to the product pages
the pagination is set as 6 per page,
reduced the amount of pages on the xl screens from 8 to 6,
Using a custom templatetag to handle if the products have queries applied
Works well with start,end, next and previous.
Having page number links doesnt work.
I have left the work in progress code commeted out for now.